cast obs to np.dtype('float32') to avoid errors while training using ray==1.6.0

+++ assistive_gym/envs/bedding_manipulation_action_series.py
@@ -341,7 +341,7 @@ class BeddingManipulationSeriesEnv(AssistiveEnv):
             output[0], output[1], output[2] = pose, all_joint_angles, all_pos_orient
             return output
 
-        return pose
+        return pose.astype('float32')
 
     def set_seed_val(self, seed = 1001):
         if seed != self.seed_val:
